parametrize simple test scenarios
