+++ tests/scenarios/test_upgrade_scenario.py
@@ -42,10 +42,10 @@ from tests.helpers import (
     assert_app_updated,
     assert_upgrade_tester_deletes_app,
 )
-from tests.scenarios.test_simple_scenarios import (
-    patch_pytest_test_runner,
-    assert_prepare_pytest_test_environment,
+from tests.scenarios.executors.pytest import (
     assert_run_pytest,
+    assert_prepare_pytest_test_environment,
+    patch_pytest_test_runner,
 )
